+++ engine/calculators/source/unit_tests/MagicalDamageCalculator_test.cpp
@@ -48,8 +48,8 @@ TEST(Engine_Calculators_MagicalDamageCalculator, base_damage_no_bonuses)
   MagicalDamageCalculator mdc(PhaseOfMoonType::PHASE_OF_MOON_NULL);
   Damage d = mdc.calculate_base_damage_object(creature);
 
-  EXPECT_EQ(3, d.get_num_dice());
-  EXPECT_EQ(4, d.get_dice_sides());
+  EXPECT_EQ(static_cast<uint>(3), d.get_num_dice());
+  EXPECT_EQ(static_cast<uint>(4), d.get_dice_sides());
   EXPECT_EQ(0, d.get_modifier());
 
   tear_down_spells();
@@ -63,8 +63,8 @@ TEST(Engine_Calculators_MagicalDamageCalculator, base_damage_spell_bonus_compone
   MagicalDamageCalculator mdc(PhaseOfMoonType::PHASE_OF_MOON_NULL);
   Damage d = mdc.calculate_base_damage_with_bonuses_or_penalties(creature);
 
-  EXPECT_EQ(3, d.get_num_dice());
-  EXPECT_EQ(4, d.get_dice_sides());
+  EXPECT_EQ(static_cast<uint>(3), d.get_num_dice());
+  EXPECT_EQ(static_cast<uint>(4), d.get_dice_sides());
   EXPECT_EQ(0, d.get_modifier());
 
   IndividualSpellKnowledge isk;
@@ -72,8 +72,8 @@ TEST(Engine_Calculators_MagicalDamageCalculator, base_damage_spell_bonus_compone
   creature->get_spell_knowledge_ref().set_spell_knowledge(spell_id, isk);
   d = mdc.calculate_base_damage_with_bonuses_or_penalties(creature);
 
-  EXPECT_EQ(3, d.get_num_dice());
-  EXPECT_EQ(4, d.get_dice_sides());
+  EXPECT_EQ(static_cast<uint>(3), d.get_num_dice());
+  EXPECT_EQ(static_cast<uint>(4), d.get_dice_sides());
   EXPECT_EQ(23, d.get_modifier());
 
   tear_down_spells();
@@ -88,16 +88,16 @@ TEST(Engine_Calculators_MagicalDamageCalculator, base_damage_skill_component)
   MagicalDamageCalculator mdc(PhaseOfMoonType::PHASE_OF_MOON_NULL);
   Damage d = mdc.calculate_base_damage_with_bonuses_or_penalties(creature);
 
-  EXPECT_EQ(3, d.get_num_dice());
-  EXPECT_EQ(4, d.get_dice_sides());
+  EXPECT_EQ(static_cast<uint>(3), d.get_num_dice());
+  EXPECT_EQ(static_cast<uint>(4), d.get_dice_sides());
   EXPECT_EQ(2, d.get_modifier());
 
   creature->get_skills().set_value(SkillType::SKILL_MAGIC_CANTRIPS, 75);
 
   d = mdc.calculate_base_damage_with_bonuses_or_penalties(creature);
 
-  EXPECT_EQ(3, d.get_num_dice());
-  EXPECT_EQ(4, d.get_dice_sides());
+  EXPECT_EQ(static_cast<uint>(3), d.get_num_dice());
+  EXPECT_EQ(static_cast<uint>(4), d.get_dice_sides());
   EXPECT_EQ(9, d.get_modifier());
 
   tear_down_spells();
@@ -113,8 +113,8 @@ TEST(Engine_Calculators_MagicalDamageCalculator, base_damage_stat_component)
   MagicalDamageCalculator mdc(PhaseOfMoonType::PHASE_OF_MOON_NULL);
   Damage d = mdc.calculate_base_damage_with_bonuses_or_penalties(creature);
 
-  EXPECT_EQ(3, d.get_num_dice());
-  EXPECT_EQ(4, d.get_dice_sides());
+  EXPECT_EQ(static_cast<uint>(3), d.get_num_dice());
+  EXPECT_EQ(static_cast<uint>(4), d.get_dice_sides());
   EXPECT_EQ(3, d.get_modifier());
 
   creature->set_intelligence(24);
@@ -122,8 +122,8 @@ TEST(Engine_Calculators_MagicalDamageCalculator, base_damage_stat_component)
 
   d = mdc.calculate_base_damage_with_bonuses_or_penalties(creature);
 
-  EXPECT_EQ(3, d.get_num_dice());
-  EXPECT_EQ(4, d.get_dice_sides());
+  EXPECT_EQ(static_cast<uint>(3), d.get_num_dice());
+  EXPECT_EQ(static_cast<uint>(4), d.get_dice_sides());
   EXPECT_EQ(5, d.get_modifier());
 
   tear_down_spells();
